+++ src/main/java/pl/tw/ksiegowosc/dto/MeritCreateInvoiceRequest.java
@@ -3,8 +3,10 @@ package pl.tw.ksiegowosc.dto;
 import java.math.BigDecimal;
 import java.util.List;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+@JsonInclude(JsonInclude.Include.NON_NULL)
 public record MeritCreateInvoiceRequest(
         @JsonProperty("Customer") MeritCreateInvoiceCustomer customer,
         @JsonProperty("AccountingDoc") int accountingDoc,